Projekte angepasst

+++ Twee2Z/CodeGen/Text/TextHelper.cs
@@ -120,7 +120,7 @@ namespace Twee2Z.CodeGen.Text
         {
             IList<UInt16> output = new List<UInt16>();
             int i = 0;
-            byte zCharCount = 0;
+            short zCharCount = 0;
 
             while (i < input.Length)
             {
@@ -171,7 +171,7 @@ namespace Twee2Z.CodeGen.Text
             return output.ToArray();
         }
 
-        private static void AddCharToArray(ref IList<UInt16> zCharList, byte number, byte table, ref byte zCharCount)
+        private static void AddCharToArray(ref IList<UInt16> zCharList, byte number, byte table, ref short zCharCount)
         {
             if (number >= 32)
                 throw new ArgumentOutOfRangeException("number", "A ZChar must have a value between 0 and 31.");
@@ -210,7 +210,7 @@ namespace Twee2Z.CodeGen.Text
             }
         }
 
-        private static void AddCharToWord(ref UInt16 word, byte number, ref byte zCharCount)
+        private static void AddCharToWord(ref UInt16 word, byte number, ref short zCharCount)
         {
             switch (zCharCount % 3)
             {
@@ -228,7 +228,7 @@ namespace Twee2Z.CodeGen.Text
             zCharCount++;
         }
 
-        private static void FinishArray(ref IList<UInt16> zCharList, ref byte zCharCount)
+        private static void FinishArray(ref IList<UInt16> zCharList, ref short zCharCount)
         {
             UInt16 word = zCharList.Last();
